# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /register:
    x-swagger-router-controller: hello_world
    get:
      description: register the user
      # used as the method name of the controller
      operationId: register
      parameters:
        - name: phonenumber
          in: query
          description: The name of the person to whom to say hello
          required: true
          type: string
        - name: password
          in: query
          description: The name of the person to whom to say hello
          required: true
          type: string
        - name: username
          in: query
          description: The name of the person to whom to say hello
          required: true
          type: string
        - name: venmoid
          in: query
          description: The name of the person to whom to say hello
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/RegisterResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /login:
    # binds a127 app logic to a route
    x-swagger-router-controller: hello_world
    get:
      description: Returns 'Hello' to the caller
      # used as the method name of the controller
      operationId: login
      parameters:
        - name: phonenumber
          in: query
          description: The name of the person to whom to say hello
          required: true
          type: string
        - name: password
          in: query
          description: The name of the person to whom to say hello
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/LoginResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /sendTask:
    # binds a127 app logic to a route
    x-swagger-router-controller: hello_world
    get:
      description: Returns 'Hello' to the caller
      # used as the method name of the controller
      operationId: sendtask
      parameters:
         - name: phonenumber
           in: query
           description: The name of the person to whom to say hello
           required: true
           type: string
         - name: timestamp
           in: query
           description: The name of the person to whom to say hello
           required: true
           type: string
         - name: longitude
           in: query
           description: The name of the person to whom to say hello
           required: true
           type: number
         - name: latitude
           in: query
           description: The name of the person to whom to say hello
           required: true
           type: number
         - name: description
           in: query
           description: The name of the person to whom to say hello
           required: true
           type: string
         - name: price
           in: query
           description: The name of the person to whom to say hello
           required: true
           type: number
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/SendTaskResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /updateTaskList:
    # binds a127 app logic to a route
    x-swagger-router-controller: hello_world
    get:
      description: Returns 'Hello' to the caller
      # used as the method name of the controller
      operationId: updatetasklist
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/UpdateResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /updateLocation:
      # binds a127 app logic to a route
      x-swagger-router-controller: hello_world
      get:
        description: Returns 'Hello' to the caller
        # used as the method name of the controller
        operationId: updatelocation
        parameters:
           - name: phonenumber
             in: query
             description: The name of the person to whom to say hello
             required: true
             type: string
           - name: longitude
             in: query
             description: The name of the person to whom to say hello
             required: true
             type: number
           - name: latitude
             in: query
             description: The name of the person to whom to say hello
             required: true
             type: number
        responses:
          "200":
            description: Success
            schema:
              # a pointer to a definition
              $ref: "#/definitions/UpdateLocationResponse"
          # responses may fall through to errors
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
  /receiveTask:
          # binds a127 app logic to a route
      x-swagger-router-controller: hello_world
      get:
        description: Returns 'Hello' to the caller
        # used as the method name of the controller
        operationId: receivetask
        parameters:
           - name: phonenumber
             in: query
             description: The name of the person to whom to say hello
             required: true
             type: string
           - name: timestamp
             in: query
             description: The name of the person to whom to say hello
             required: true
             type: string
           - name: receiver
             in: query
             description: The name of the person to whom to say hello
             required: true
             type: string
        responses:
          "200":
            description: Success
            schema:
              # a pointer to a definition
              $ref: "#/definitions/ReceiveTaskResponse"
          # responses may fall through to errors
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
  /confirmTask:
        # binds a127 app logic to a route
        x-swagger-router-controller: hello_world
        get:
          description: Returns 'Hello' to the caller
          # used as the method name of the controller
          operationId: confirmtask
          parameters:
             - name: phonenumber
               in: query
               description: The name of the person to whom to say hello
               required: true
               type: string
             - name: timestamp
               in: query
               description: The name of the person to whom to say hello
               required: true
               type: string
          responses:
            "200":
              description: Success
              schema:
                # a pointer to a definition
                $ref: "#/definitions/ConfirmTaskResponse"
            # responses may fall through to errors
            default:
              description: Error
              schema:
                $ref: "#/definitions/ErrorResponse"
  /pollingLocation:
    # binds a127 app logic to a route
    x-swagger-router-controller: hello_world
    get:
      description: Returns 'Hello' to the caller
      # used as the method name of the controller
      operationId: pollinglocation
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/OpenResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
# complex objects have schema definitions
definitions:
  OpenResponse:
      properties: {}
  ReceiveTaskResponse:
      required:
        - status
      properties:
        status:
          type: string
  ConfirmTaskResponse:
      required:
        - status
      properties:
        status:
          type: string
  PollingResponse:
      required:
        - locationList
      properties:
        locationList:
          type: array
          items:
            $ref: '#/definitions/Location'
  Location:
      required:
        - phonenumber
        - longitude
        - latitude
      properties:
        phonenumber:
          type: string
        longitude:
          type: number
        latitude:
          type: number
  UpdateLocationResponse:
      required:
        - status
      properties:
        status:
          type: string
  UpdateResponse:
      required:
        - taskList
      properties:
        taskList:
          type: array
          items:
            $ref: '#/definitions/Task'
  Task:
      required:
        - phonenumber
        - timestamp
        - longitude
        - latitude
        - description
        - price
        - receiver
        - status
      properties:
        phonenumber:
          type: string
        timestamp:
          type: string
        longitude:
          type: number
        latitude:
          type: number
        description:
          type: string
        price:
          type: number
        receiver:
          type: string
        status:
          type: string
  SendTaskResponse:
      required:
        - registerResult
      properties:
        registerResult:
          type: string
  RegisterResponse:
      required:
        - registerResult
      properties:
        registerResult:
          type: string
  LoginResponse:
    required:
      - status
    properties:
      status:
        type: string
      username:
        type: string
      password:
        type: string
      phonenumber:
        type: string
      venmoid:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string